add server interface
