+++ pkg/transport/transport.go
@@ -12,3 +12,8 @@ type MerlinResponse struct {
 type MerlinCommClient interface {
 	Do(b io.Reader) (MerlinResponse, error)
 }
+
+type MerlinServerClient interface {
+	Run() error
+	RegisterHandler(func(w io.Writer, r io.Reader)) MerlinServerClient
+}
